HPcorrs

+++ StRoot/StSvtClusterMaker/StSvtAnalysedHybridClusters.hh
@@ -1,6 +1,6 @@
 /***************************************************************************
  *
- * $Id: StSvtAnalysedHybridClusters.hh,v 1.2 2000/08/24 04:27:56 caines Exp $
+ * $Id: StSvtAnalysedHybridClusters.hh,v 1.3 2001/04/25 18:23:21 perev Exp $
  *
  * Author: Selemon Bekele
  ***************************************************************************
@@ -10,6 +10,9 @@
  ***************************************************************************
  *
  * $Log: StSvtAnalysedHybridClusters.hh,v $
+ * Revision 1.3  2001/04/25 18:23:21  perev
+ * HPcorrs
+ *
  * Revision 1.2  2000/08/24 04:27:56  caines
  * Fixed casting warnings so compiles without errors on linux
  *
@@ -67,7 +70,7 @@ class StSvtAnalysedHybridClusters : public StSvtHybridObject
  private:
     
     int mNumOfHits;
-    ulong mHardWarePosition;
+    unsigned int mHardWarePosition;
     StSvtHitData* mSvtHitData;     //!
     StSvtHit* mSvtHit;          //!   
     StThreeVector<double>* mPos;        //!
